tasks/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-slack-bot-tasks"
version = "1.2.6"
description = "APScheduler WebUI for AI Slack Bot - Scheduled task management"
authors = [
    {name = "Travis Frisinger", email = "travis@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"

dependencies = [
    # Core scheduling
    "apscheduler>=3.10.4",

    # Web framework (migrated from Flask to FastAPI for consistency)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",  # ASGI server for production
    "jinja2>=3.1.0",  # Template rendering for OAuth pages
    "itsdangerous>=2.1.0",  # Required by Starlette SessionMiddleware

    # Time zones
    "tzlocal>=5.2",
    "pytz>=2023.3",

    # Configuration & environment
    "python-dotenv>=1.0.0",
    "pydantic>=1.9.0",
    "pydantic-settings>=2.0.0",

    # HTTP & requests for API calls
    "requests>=2.28.0",
    "aiohttp>=3.8.0",
    "httpx>=0.27.0",  # For SEC Edgar API
    "beautifulsoup4>=4.12.0",  # For parsing SEC HTML filings
    "edgartools>=2.0.0",  # For parsing SEC filings (10-K, 10-Q, 8-K, etc.)

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "pymysql>=1.1.0",
    "cryptography>=41.0.0",  # Required for MySQL auth methods
    "alembic>=1.13.0",

    # Google Drive API (for scheduled ingestion)
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",

    # Document processing (for Google Drive ingestion)
    "PyMuPDF>=1.23.0",  # PDF processing (fitz)
    "python-docx>=1.1.0",  # Word document processing
    "openpyxl>=3.1.0",  # Excel spreadsheet processing
    "python-pptx>=0.6.23",  # PowerPoint presentation processing

    # Slack SDK (for user import)
    "slack-sdk>=3.35.0",

    # GraphQL client (for Metric.ai sync)
    "gql[requests]>=3.4.0",

    # Vector DB and Embeddings (for Metric.ai sync)
    "qdrant-client>=1.7.0",
    "openai>=1.0.0",

    # Logging
    "loguru>=0.7.0",

    # JWT for Portal API authentication
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
# Test dependencies
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "fastapi>=0.115.0",  # Required by test factories
]

# Development tools
dev = [
    "ruff==0.14.0",
    "pyright==1.1.390",
    "pre-commit==4.0.1",
]

[project.scripts]
tasks-server = "app:main"

# Package discovery configuration
[tool.setuptools]
packages = ["config", "jobs", "services", "models", "tests"]

# Ruff configuration
[tool.ruff]
target-version = "py311"
line-length = 88
src = ["."]
exclude = [
    ".bzr", ".direnv", ".eggs", ".git", ".git-rewrite", ".hg",
    ".mypy_cache", ".nox", ".pants.d", ".pytype", ".ruff_cache",
    ".svn", ".tox", ".venv", "__pypackages__", "_build",
    "buck-out", "build", "dist", "node_modules", "venv", "htmlcov",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "PL",   # pylint
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long, handled by ruff format
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "PLR0913", # too many arguments
    "PLR0912", # too many branches
    "PLR0915", # too many statements
    "PLR2004", # magic value used in comparison
    "PLW0603", # Using the global statement is discouraged (needed for Flask app factory)
    "PLC0415", # Allow imports inside functions for optional dependencies
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "B011", "F811"]  # Allow tests to use asserts, etc.

[tool.ruff.lint.isort]
known-first-party = ["."]

# Ruff formatting configuration
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Pyright configuration
[tool.pyright]
pythonVersion = "3.11"
include = ["."]
exclude = ["tests", "**/__pycache__", "**/node_modules"]
typeCheckingMode = "standard"

# Coverage configuration (centralized)
[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
    "*/migrations/*",
    "setup.py",
    "app.py",  # FastAPI entry point - tested via API endpoints
    "run_metric_sync.py",  # CLI script
    # Optional Google Drive integration (complex external service integration)
    "api/gdrive.py",  # Optional feature - Google Drive OAuth and ingestion APIs
    "jobs/gdrive_ingest.py",  # Optional feature - Google Drive ingestion job
    "services/gdrive/*",  # Optional Google Drive services
    # Optional integrations
    "jobs/portal_sync.py",  # Optional feature - Portal sync job
    "jobs/metric_sync.py",  # Optional feature - Metrics sync job
    "services/portal_client.py",  # Optional feature - Portal API client
    "services/metric_client.py",  # Optional feature - Metrics API client
    "jobs/system/*",  # System job implementations tested via job_registry tests
]

[tool.coverage.report]
fail_under = 70
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_paths = ["."]
asyncio_mode = "auto"
addopts = [
    "--verbose",
    "--cov=.",
    "--cov-fail-under=70",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
]
markers = [
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
